requirements.txt
cx-Freeze==5.1.1
idna==2.8
Pillow==5.3.0
